// Client/bindings/FieldEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FieldEventValue } from "./FieldEventValue";

export type FieldEvent = {
  id: number;
  time: string;
  field_id: number;
  type_id: number;
  type_name: string;
  note: string | null;
  values: { [key: string]: FieldEventValue };
};